requirements.txt
fastapi
uvicorn
celery
pymongo
python-multipart
pillow
requests
pandas
redis
aiofiles
python-dotenv
cryptography
pyOpenSSL
